Address minor feedback
